# docs/swagger.yaml
definitions:
...
  models.Seller:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Sku:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      seller_id:
        type: string
      sku_code:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Tenant:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /sellers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Seller'
            type: array
      summary: Get all sellers
      tags:
      - Sellers
    post:
      consumes:
      - application/json
      parameters:
      - description: Seller to create
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/models.Seller'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Seller'
      summary: Create a new seller
      tags:
      - Sellers
  /sellers/{id}:
    delete:
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
      summary: Delete seller by ID
      tags:
      - Sellers
    get:
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
      summary: Get seller by ID
      tags:
      - Sellers
    put:
      consumes:
      - application/json
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated seller
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/models.Seller'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
      summary: Update seller by ID
      tags:
      - Sellers
  /skus:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Filter by Seller ID
        in: query
        name: seller_id
        type: string
      - collectionFormat: csv
        description: Filter by multiple SKU codes (repeat param)
        in: query
        items:
          type: string
        name: sku_codes
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Sku'
            type: array
      summary: Get all SKUs (with optional filters)
      tags:
      - SKUs
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: SKU to create
        in: body
        name: sku
        required: true
        schema:
          $ref: '#/definitions/models.Sku'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Sku'
      summary: Create a new SKU
      tags:
      - SKUs
  /skus/{id}:
    delete:
      parameters:
      - description: SKU ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sku'
      summary: Delete SKU by ID
      tags:
      - SKUs
    get:
      parameters:
      - description: SKU ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sku'
      summary: Get SKU by ID
      tags:
      - SKUs
    put:
      consumes:
      - application/json
      parameters:
      - description: SKU ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Updated SKU
        in: body
        name: sku
        required: true
        schema:
          $ref: '#/definitions/models.Sku'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sku'
      summary: Update SKU by ID
      tags:
      - SKUs
  /tenants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tenant'
            type: array
      summary: Get all tenants
      tags:
      - Tenants
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant to create
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tenant'
      summary: Create a new tenant
      tags:
      - Tenants
  /tenants/{id}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
      summary: Delete tenant by ID
      tags:
      - Tenants
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tenant by ID
      tags:
      - Tenants
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant data
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
      summary: Update tenant by ID
      tags:
      - Tenants
...
swagger: "2.0"